% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{grepdt}
\alias{grepdt}
\title{grepdt}
\usage{
grepdt(dt, str)
}
\arguments{
\item{dt}{Data frame}

\item{str}{Cadena de caracteres que queremos encontrar o regex.}
}
\value{
vector de caracteres
}
\description{
Esta funcion busca todas las columnas cuyo nombre coincida con el str dado.
}
