% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{types}
\alias{types}
\title{types}
\usage{
types(dt, sort = T)
}
\arguments{
\item{dt}{Data frame del que queremos saber los tipo de datos.}

\item{sort}{TRUE si queremos que lo devuelva ordenado por frecuencia.}
}
\value{
data.table data.frame
}
\description{
Función que devuelve el tipo de dato en cada columna del data.frame
}
