% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{value_counts}
\alias{value_counts}
\title{value_counts}
\usage{
value_counts(column, sort = T)
}
\arguments{
\item{column}{Columna de data.frame.}

\item{sort}{TRUE si queremos que lo devuelva ordenado por frecuencia.}
}
\value{
data.table data.frame
}
\description{
Cuenta valores distintos en una columna de data.frame
}
